% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.weight.R
\name{var.weight}
\alias{var.weight}
\title{Maximum variance weights}
\usage{
var.weight(X)
}
\arguments{
\item{X}{n by p maxtrix containing observations of p biomarkers of n subjects.}
}
\description{
var.weight produces a set of weights that maximizes the total weighted variance of the distribution of different biomarkers within each subject.
}
\examples{
library(MASS)
# a three biomarkers dataset generated from independent normal(0,1)
X = mvrnorm(n = 100, mu=rep(0,3), Sigma=diag(3), tol = 1e-6, empirical = FALSE, EISPACK = FALSE)
var.weight(X)
}
\keyword{weighting}
